Release v4.1.0 — Sprint 17 SCPI parts fractionnées (#11)
* PROJECT_PLAN met à jour vers v3.2 — état v3.0.1 et workflow git

- Statut v3.0.0 → v3.0.1 (CI vert, Play Store en attente)
- F7 Auth : email/password retiré (PIN + biométrie uniquement)
- Room v4 → v7, migrations 1→7, modèles enrichis (CustomSubCategory, MonthlyVersement)
- UseCases 37+ → 40+, tests 110 → 123
- CI : tests unitaires ✅, couverture domain/ ✅, protection main ✅
- Workflow git : develop + feature/xxx + PR obligatoire sur main documenté
- Sprint 12 chantiers A/B/C ajoutés (auth, édition, catégories, locale, biométrie, devise, rename)
- Sprint 12b CI-FIX ajouté, Sprint 13 ajouté comme "À venir"
- Risques : "5e migration" → "8e migration" (Room v7 stable)


* Sprint 13 — Kover, Crashlytics, taux de change Frankfurter (#3)

* PROJECT_PLAN passe en v3.3 — Sprint 14 backlog et convention versioning


* Sprint 13 — Kover corrigé, Crashlytics, taux de change Frankfurter

- TECH-01 : bloc kover déplacé au niveau projet (hors android{}), seuil 80%
  domain/, rapport HTML uploadé en CI artefact
- TECH-02 : Firebase Crashlytics intégré (google-services.json.example fourni,
  collection désactivée en debug, BuildConfig.CRASHLYTICS_ENABLED)
- FEAT-04 : API Frankfurter EUR/USD/XOF — ExchangeRates domain, repository
  avec cache DataStore 1h, NetworkModule Hilt, section Paramètres
- AndroidManifest : permission INTERNET ajoutée
- SettingsViewModelTest : mock ucRates ajouté (124 tests CI verts)


---------


* Sprint 14 — IME clavier, analyse récurrence, donut interactif (#5)

* UX-02 — couverture IME complète sur tous les écrans de saisie

Chaque champ éditables expose désormais le bon bouton d'action :
- ImeAction.Next + moveFocus pour les champs intermédiaires
- ImeAction.Done + clearFocus pour le dernier champ du formulaire
- ImeAction.Search sur la barre de recherche des dépenses
- FocusRequester pour sauter les champs readOnly (ExpenseSheet)
- imePadding() ajouté sur les AlertDialog manquants (Budget, Settings)

Screens traités : Expenses, Budget, Debts, Investments (6 sheets),
Savings (2 sheets + dialog), Settings.


* PROJECT_PLAN passe en v3.4 — ANALYSE-01 et UX-02 Sprint 14

ANALYSE-01 : NO-GO intégration RecurringExpenseTracker (DennisBauer)
— GPL-3.0 incompatible, stack KMP vs Android-natif.
Concepts retenus pour backlog v4 : récurrences enrichies + vue
prochains paiements (FEAT-RECUR).
UX-02 marqué livré dans le tableau Sprint 14.


* FEAT-BUDGET-INT — graphique donut interactif avec drill-down AUTRE et navigation revenus/dépenses

- BudgetScreen : camembert cliquable par catégorie, drill-down sous-catégories AUTRE, liens revenus/dépenses vers ExpensesScreen avec filtre pré-rempli
- BudgetViewModel : ajoute getCustomSubCategories au combine, expose customSubCategories dans Success
- ExpensesViewModel : initialise le filtre depuis SavedStateHandle (args de navigation category + type)
- DibitaraNavGraph : route Expenses avec args optionnels, BudgetScreen reçoit onNavigateToExpenses
- Tests : constructeurs mis à jour (getCustomSubCategories + SavedStateHandle)

* Sprint 15 — Suggestions de saisie rapide (FEAT-SUGGEST) (#7)

* Met à jour PROJECT_PLAN.md v3.6 — aligne état réel Sprints 13 + 14

Sprints 13 et 14 marqués terminés, stack technique corrigée
(Kover, Crashlytics, Frankfurter en ✅), version 3.1.0, F8/F9 → backlog V4.
---
* Définit Sprint 15 — FEAT-SUGGEST suggestions de saisie rapide
---
GetTransactionSuggestionsUseCase + TransactionSuggestion + chips UI,
5-8h, v3.2.0, sans migration Room.
---
* Sprint 15 : Suggestions de saisie rapide basées sur l'historique récent

GetTransactionSuggestionsUseCase agrège les transactions des 30 derniers
jours (seuil ≥ 2 occurrences, top 5 par fréquence). ExpensesViewModel expose
un StateFlow<List<TransactionSuggestion>>. Les chips apparaissent sous le
champ Note et pré-remplissent tous les champs en un tap. 7 tests ajoutés.
---------

* Met à jour PROJECT_PLAN.md v3.8 — Sprint 15 marqué terminé (v3.2.0)


* FEAT-RECUR — récurrences hebdo/annuelles + vue prochains paiements (#8)

- RecurrenceFrequency enum (WEEKLY / MONTHLY / YEARLY)
- Transaction : +recurrenceFrequency, +firstPaymentDate, +endDate
- Room v7→v8 : migration 3 colonnes, backfill MONTHLY sur existants
- GenerateRecurringUseCase remplace GenerateMonthlyRecurringUseCase :
  rattrapage de toutes les occurrences manquantes, respect endDate
- GetUpcomingPaymentsUseCase : prochaines N échéances triées par date
- DashboardViewModel + DashboardScreen : carte "Prochains paiements"
- ExpensesSheet : dropdown fréquence + sélecteur date de fin
- ExpensesViewModel : propagation recurrenceFrequency + endDate
- Tests : GenerateRecurringUseCase couvre les 3 fréquences


* Version passe à 4.0.0 (versionCode 11) — FEAT-RECUR Room v7→v8

Sprint 16 (FEAT-RECUR) marqué terminé dans PROJECT_PLAN.md v3.9 :
récurrences enrichies WEEKLY/MONTHLY/YEARLY, vue prochains paiements,
migration Room v7→v8, versionCode 11.


* Paramètres contrôlent l'affichage de Prochains paiements

Toggle dans la section Tableau de bord (Paramètres) pour afficher
ou masquer la carte "Prochains paiements" sur le Dashboard, même
comportement que le toggle Rapport mensuel.


* Met à jour PROJECT_PLAN.md v4.0 — corrections pré-déploiement UX-TOGGLE-PP


* Ajoute protocole de protection des données Room (§8.1)

Checklist pré-migration, alertes bloquantes et tableau des opérations
à risque — appliqué avant chaque sprint qui modifie le schéma Room.


* Tests couvrent UpdateAfficherProchainsPaiementsUseCase — seuil Kover 80%


* SCPI accepte les parts fractionnées — Room v8→v9

sharesCount passe de Int à Double (REAL en SQLite) pour permettre
des valeurs comme 2,2 parts. Migration v8→v9 recrée la table
scpi_investments en préservant les données existantes.


* Version passe à 4.1.0 (versionCode 12) — Sprint 17 SCPI


---------

+++ app/src/main/java/com/dibitara/app/domain/model/Investment.kt
@@ -13,14 +13,14 @@ data class RealEstateAsset(
 data class ScpiInvestment(
     val id: Long = 0,
     val label: String,
-    val sharesCount: Int,
+    val sharesCount: Double,
     val shareValueCents: Long,
     val monthlyContributionCents: Long,
     val currency: Currency,
     val updatedAt: LocalDate
 ) {
-    // Valeur totale = nombre de parts × valeur unitaire
-    val totalValueCents: Long get() = sharesCount.toLong() * shareValueCents
+    // Valeur totale = nombre de parts (peut être fractionnaire, ex : 2,2) × valeur unitaire
+    val totalValueCents: Long get() = (sharesCount * shareValueCents).toLong()
 }
 
 data class AirbnbRental(
